Handle modules somewhat better
The weird thing is still that we can have duplicate names that we don't
properly namespace. Maybe we should embrace the module system and track
which things are named in a given scope - then local open would make
more sense, and we can emit more focused errors when there are
duplicates.