Successfully write some functions! (See description)
The output is confusing because the name section is not being modified.
This is annoying but we can solve that another day. The imports and
functions are being correctly preserved though.

Types will need to be relocated eventually, if we do in fact relocate
types at all (and I think it would be nice).

Function bodies are not being written at all. We need to iterate
instructions and do any necessary relocations on them.

Function locals are not being written at all.